moved client function docstrings from pusher.py to *_client.py
